+++ packages/web/app/api/queries/useAdminPermissionsQuery.js
@@ -1,0 +1,12 @@
+import { useQuery } from '@tanstack/react-query';
+import { useApi } from '../useApi';
+
+export const useAdminPermissionsQuery = (rolesQueryParam, options = {}) => {
+  const api = useApi();
+
+  return useQuery(
+    ['adminPermissions', rolesQueryParam],
+    () => api.get('admin/permissions', { roles: rolesQueryParam }),
+    options,
+  );
+};
